Move rewards checks into its own package
